Integrated previously written unit tests.
